+++ python/genvarloader/_dataset/_genotypes.py
@@ -4,12 +4,13 @@ from typing import Optional, Tuple
 import numba as nb
 import numpy as np
 from attrs import define
+from genoray._svar import SparseGenotypes
 from numpy.typing import NDArray
 
 from .._types import ListIdx
 from .._utils import _lengths_to_offsets
 
-__all__ = []
+__all__ = ["SparseGenotypes"]
 
 
 @define
@@ -113,185 +114,156 @@ def get_haplotype_region_ilens(
     return r_ilens
 
 
-@define
-class SparseGenotypes:
-    """Sparse genotypes corresponding to distinct regions. In this format, genotypes are stored as a ragged 3D array where each
-    sample, ploid, and region may have a different number of variants, since unknown and REF genotypes are not stored. The
-    variant indices are aligned to the genotypes. Physically, the genotypes and variant indices are stored as 1D arrays.
-    Then, each sample and region's info can be sliced out using the offsets:
-    >>> i = np.ravel_multi_index((r, p, s), (n_regions, ploidy, n_samples))
-    >>> variant_idxs[offsets[i]:offsets[i+1]]
+# class SparseGenotypes(Ragged[np.int32]):
+#     """Sparse genotypes corresponding to distinct regions. In this format, genotypes are stored as a ragged 3D array where each
+#     sample, ploid, and region may have a different number of variants, since unknown and REF genotypes are not stored. The
+#     variant indices are aligned to the genotypes. Physically, the genotypes and variant indices are stored as 1D arrays.
+#     Then, each sample and region's info can be sliced out using the offsets:
+#     >>> i = np.ravel_multi_index((r, p, s), (n_regions, ploidy, n_samples))
+#     >>> variant_idxs[offsets[i]:offsets[i+1]]
 
-    Attributes
-    ----------
-    variant_idxs : NDArray[np.int32]
-        Shape = (variants * samples * ploidy) Variant indices.
-    offsets : NDArray[np.int32]
-        Shape = (regions * samples * ploidy + 1) Offsets into genos.
-    n_samples : int
-        Number of samples.
-    ploidy : int
-        Ploidy.
-    n_regions : int
-        Number of regions.
-    """
+#     Attributes
+#     ----------
+#     variant_idxs : NDArray[np.int32]
+#         Shape = (variants * samples * ploidy) Variant indices.
+#     offsets : NDArray[np.int32]
+#         Shape = (regions * samples * ploidy + 1) Offsets into genos.
+#     n_samples : int
+#         Number of samples.
+#     ploidy : int
+#         Ploidy.
+#     n_regions : int
+#         Number of regions.
+#     """
 
-    variant_idxs: NDArray[np.int32]  # (variants * samples * ploidy)
-    offsets: NDArray[np.int64]  # (regions * samples * ploidy + 1)
-    n_regions: int
-    n_samples: int
-    ploidy: int
-    dosage: Optional[NDArray[np.float32]] = None  # (variants * samples)
+#     @property
+#     def variant_idxs(self):
+#         return self.data
 
-    @property
-    def effective_shape(self):
-        return (self.n_regions, self.n_samples, self.ploidy)
+#     @property
+#     def n_regions(self):
+#         return self.shape[0]
 
-    @classmethod
-    def empty(cls, n_regions: int, n_samples: int, ploidy: int):
-        """Create an empty sparse genotypes object."""
-        return cls(
-            np.empty(0, np.int32),
-            np.zeros(n_regions * n_samples * ploidy + 1, np.int64),
-            n_regions,
-            n_samples,
-            ploidy,
-        )
+#     @property
+#     def n_samples(self):
+#         return self.shape[1]
 
-    @property
-    def is_empty(self) -> bool:
-        return len(self.variant_idxs) == 0
+#     @property
+#     def ploidy(self):
+#         return self.shape[2]
 
-    def vars(self, region: int, sample: int, ploidy: int):
-        """Get variant indices for a given sample and region."""
-        i = np.ravel_multi_index(
-            (region, sample, ploidy), (self.n_regions, self.n_samples, self.ploidy)
-        )
-        vars = self.variant_idxs[self.offsets[i] : self.offsets[i + 1]]
-        return vars
+#     @property
+#     def effective_shape(self):
+#         return (self.n_regions, self.n_samples, self.ploidy)
 
-    def concat(*genos: "SparseGenotypes") -> "SparseGenotypes":
-        """Concatenate sparse genotypes."""
+#     @classmethod
+#     def empty(cls, n_regions: int, n_samples: int, ploidy: int):
+#         """Create an empty sparse genotypes object."""
+#         return cls.from_offsets(
+#             np.empty(0, np.int32),
+#             (n_regions, n_samples, ploidy),
+#             np.zeros(n_regions * n_samples * ploidy + 1, np.int64),
+#         )
 
-        if not all(g.n_samples == genos[0].n_samples for g in genos):
-            raise ValueError("All genotypes must have the same number of samples.")
-        if not all(g.ploidy == genos[0].ploidy for g in genos):
-            raise ValueError("All genotypes must have the same ploidy.")
+#     @property
+#     def is_empty(self) -> bool:
+#         return len(self.variant_idxs) == 0
 
-        total_n_regions = sum(g.n_regions for g in genos)
-        variant_idxs = np.concatenate([g.variant_idxs for g in genos])
-        offsets = _lengths_to_offsets(
-            np.concatenate([np.diff(g.offsets) for g in genos])
-        )
-        return SparseGenotypes(
-            variant_idxs=variant_idxs,
-            offsets=offsets,
-            n_regions=total_n_regions,
-            n_samples=genos[0].n_samples,
-            ploidy=genos[0].ploidy,
-        )
+#     def vars(self, region: int, sample: int, ploidy: int):
+#         """Get variant indices for a given sample and region."""
+#         i = np.ravel_multi_index(
+#             (region, sample, ploidy), (self.n_regions, self.n_samples, self.ploidy)
+#         )
+#         vars = self.variant_idxs[self.offsets[i] : self.offsets[i + 1]]
+#         return vars
 
-    @classmethod
-    def from_dense(
-        cls,
-        genos: NDArray[np.int8],
-        first_v_idxs: NDArray[np.int32],
-        offsets: NDArray[np.int64],
-        dosages: Optional[NDArray[np.float32]] = None,
-    ):
-        """Convert dense genotypes to sparse genotypes.
+#     @classmethod
+#     def from_dense(
+#         cls,
+#         genos: NDArray[np.int8],
+#         var_idxs: NDArray[np.integer],
+#         offsets: NDArray[np.int64],
+#     ):
+#         """Convert dense genotypes to sparse genotypes.
 
-        Parameters
-        ----------
-        genos : NDArray[np.int8]
-            Shape = (sample, ploidy, variants) Genotypes.
-        first_v_idxs : NDArray[np.uint32]
-            Shape = (regions) First variant index for each region.
-        offsets : NDArray[np.uint32]
-            Shape = (regions + 1) Offsets into genos.
-        dosages : Optional[NDArray[np.float32]]
-            Shape = (sample, ploidy, variants) Dosages.
-        """
-        n_regions = len(first_v_idxs)
-        n_samples = genos.shape[0]
-        ploidy = genos.shape[1]
-        # (s p v)
-        keep = genos == 1
-        n_per_rsp = get_n_per_rsp(keep, offsets, n_regions)
-        sparse_offsets = _lengths_to_offsets(n_per_rsp.ravel(), np.int64)
-        variant_idxs = keep_mask_to_rsp_v_idx(
-            keep, first_v_idxs, offsets, sparse_offsets, n_regions, n_samples, ploidy
-        )
-        if dosages is not None:
-            dosages = dosages[keep]
-        return cls(
-            variant_idxs=variant_idxs,
-            offsets=sparse_offsets,
-            n_regions=n_regions,
-            n_samples=n_samples,
-            ploidy=ploidy,
-            dosage=dosages,
-        )
+#         Parameters
+#         ----------
+#         genos : NDArray[np.int8]
+#             Shape = (sample ploidy ~variants) Genotypes.
+#         var_idxs : NDArray[np.uint32]
+#             Shape = (regions ~variants) variant indices for each region.
+#         offsets : NDArray[np.uint32]
+#             Shape = (regions + 1) Offsets into genotypes and var_idxs for each region.
+#         dosages : Optional[NDArray[np.float32]]
+#             Shape = (sample ploidy ~variants) Dosages.
+#         """
+#         n_regions = len(offsets) - 1
+#         n_samples = genos.shape[0]
+#         ploidy = genos.shape[1]
+#         # (s p v)
+#         keep = genos == 1
+#         n_per_rsp = get_n_per_rsp(keep, offsets, n_regions)
+#         sparse_offsets = _lengths_to_offsets(n_per_rsp.ravel(), np.int64)
+#         variant_idxs = keep_mask_to_rsp_v_idx(
+#             keep, var_idxs, offsets, sparse_offsets, n_regions, n_samples, ploidy
+#         )
+#         shape = (n_regions, n_samples, ploidy)
+#         return cls.from_offsets(variant_idxs, shape, sparse_offsets)
 
-    @classmethod
-    def from_dense_with_length(
-        cls,
-        genos: NDArray[np.int8],
-        first_v_idxs: NDArray[np.int32],
-        offsets: NDArray[np.int64],
-        ilens: NDArray[np.int32],
-        positions: NDArray[np.int32],
-        starts: NDArray[np.int32],
-        lengths: NDArray[np.int32],
-    ):
-        """Convert dense genotypes to sparse genotypes.
+#     @classmethod
+#     def from_dense_with_length(
+#         cls,
+#         genos: NDArray[np.int8],
+#         first_v_idxs: NDArray[np.int32],
+#         offsets: NDArray[np.int64],
+#         ilens: NDArray[np.int32],
+#         positions: NDArray[np.int32],
+#         starts: NDArray[np.int32],
+#         lengths: NDArray[np.int32],
+#     ):
+#         """Convert dense genotypes to sparse genotypes.
 
-        Parameters
-        ----------
-        genos : NDArray[np.int8]
-            Shape = (sample, ploidy, variants) Genotypes.
-        first_v_idxs : NDArray[np.uint32]
-            Shape = (regions) First variant index for each region.
-        offsets : NDArray[np.uint32]
-            Shape = (regions + 1) Offsets into genos.
-        ilens : NDArray[np.int32]
-            Shape = (total_variants) ILEN of all unique variants.
-        positions : NDArray[np.int32]
-            Shape = (total_variants) Positions of unique variants.
-        starts : NDArray[np.int32]
-            Shape = (regions) Start of query regions.
-        lengths : NDArray[np.int32]
-            Shape = (regions) Lengths of the output haplotypes.
-        """
-        n_regions = len(first_v_idxs)
-        n_samples = genos.shape[0]
-        ploidy = genos.shape[1]
-        # (s p v)
-        keep, min_ilens = get_keep_mask_for_length(
-            genos,
-            offsets,
-            first_v_idxs,
-            positions,
-            ilens,
-            starts,
-            lengths,
-        )
-        # (r)
-        max_ends: NDArray[np.int32] = starts + lengths - min_ilens.clip(max=0)
-        # (r s p)
-        n_per_rsp = get_n_per_rsp(keep, offsets, n_regions)
-        sparse_offsets = _lengths_to_offsets(n_per_rsp.ravel(), np.int64)
-        variant_idxs = keep_mask_to_rsp_v_idx(
-            keep, first_v_idxs, offsets, sparse_offsets, n_regions, n_samples, ploidy
-        )
-        sparse_genos = cls(
-            variant_idxs=variant_idxs,
-            offsets=sparse_offsets,
-            n_regions=n_regions,
-            n_samples=n_samples,
-            ploidy=ploidy,
-        )
-        return sparse_genos, max_ends
+#         Parameters
+#         ----------
+#         genos : NDArray[np.int8]
+#             Shape = (sample, ploidy, variants) Genotypes.
+#         first_v_idxs : NDArray[np.uint32]
+#             Shape = (regions) First variant index for each region.
+#         offsets : NDArray[np.uint32]
+#             Shape = (regions + 1) Offsets into genos.
+#         ilens : NDArray[np.int32]
+#             Shape = (total_variants) ILEN of all unique variants.
+#         positions : NDArray[np.int32]
+#             Shape = (total_variants) Positions of unique variants.
+#         starts : NDArray[np.int32]
+#             Shape = (regions) Start of query regions.
+#         lengths : NDArray[np.int32]
+#             Shape = (regions) Lengths of the output haplotypes.
+#         """
+#         n_regions = len(first_v_idxs)
+#         n_samples = genos.shape[0]
+#         ploidy = genos.shape[1]
+#         # (s p v)
+#         keep, min_ilens = get_keep_mask_for_length(
+#             genos,
+#             offsets,
+#             first_v_idxs,
+#             positions,
+#             ilens,
+#             starts,
+#             lengths,
+#         )
+#         # (r)
+#         max_ends: NDArray[np.int32] = starts + lengths - min_ilens.clip(max=0)
+#         # (r s p)
+#         n_per_rsp = get_n_per_rsp(keep, offsets, n_regions)
+#         sparse_offsets = _lengths_to_offsets(n_per_rsp.ravel(), np.int64)
+#         variant_idxs = keep_mask_to_rsp_v_idx(
+#             keep, first_v_idxs, offsets, sparse_offsets, n_regions, n_samples, ploidy
+#         )
+#         shape = (n_regions, n_samples, ploidy)
+#         sparse_genos = cls.from_offsets(variant_idxs, shape, sparse_offsets)
+#         return sparse_genos, max_ends
 
 
 @nb.njit(parallel=True, nogil=True, cache=True)
@@ -398,8 +370,8 @@ def get_n_per_rsp(keep: NDArray[np.bool_], offsets: NDArray[np.int64], n_regions
 
 @nb.njit(parallel=True, nogil=True, cache=True)
 def keep_mask_to_rsp_v_idx(
-    keep: NDArray[np.bool_],  # (s p v)
-    first_v_idxs: NDArray[np.int32],  # (r)
+    keep: NDArray[np.bool_],  # (s p ~v)
+    var_idxs: NDArray[np.integer],  # (r ~v)
     offsets: NDArray[np.int64],  # (r + 1)
     sparse_offsets: NDArray[np.int64],  # (r*s*p + 1)
     n_regions,
@@ -408,18 +380,19 @@ def keep_mask_to_rsp_v_idx(
 ):
     variant_idxs = np.empty(sparse_offsets[-1], np.int32)
     for r in nb.prange(n_regions):
-        fvi = first_v_idxs[r]
         o_s, o_e = offsets[r], offsets[r + 1]
         n_variants = o_e - o_s
         if n_variants == 0:
             continue
+        # (v)
+        region_var_idxs = var_idxs[o_s:o_e]
         for s in nb.prange(n_samples):
             for p in nb.prange(ploidy):
                 out_start = sparse_offsets[r * n_samples * ploidy + s * ploidy + p]
                 out_step = 0
                 for v in range(n_variants):
                     if keep[s, p, o_s + v]:
-                        variant_idxs[out_start + out_step] = fvi + v
+                        variant_idxs[out_start + out_step] = region_var_idxs[v]
                         out_step += 1
     return variant_idxs
 
@@ -448,7 +421,7 @@ class SparseSomaticGenotypes:
         Number of regions.
     """
 
-    variant_idxs: NDArray[np.int32]  # (variants * samples)
+    data: NDArray[np.int32]  # (variants * samples)
     ccfs: NDArray[np.float32]  # (variants * samples)
     offsets: NDArray[np.int64]  # (regions * samples + 1)
     n_regions: int
@@ -456,7 +429,7 @@ class SparseSomaticGenotypes:
     ploidy = 1
 
     @property
-    def effective_shape(self):
+    def shape(self):
         """Effective shape of the sparse genotypes (n_regions, n_samples, ploidy)
         where ploidy is always represented as 1. The ploidy is treated as 1 to be consistent with."""
         return (self.n_regions, self.n_samples, self.ploidy)
@@ -474,12 +447,12 @@ class SparseSomaticGenotypes:
 
     @property
     def is_empty(self) -> bool:
-        return len(self.variant_idxs) == 0
+        return len(self.data) == 0
 
     def vars(self, region: int, sample: int):
         """Get variant indices for a given sample and region."""
         i = np.ravel_multi_index((region, sample), (self.n_regions, self.n_samples))
-        vars = self.variant_idxs[self.offsets[i] : self.offsets[i + 1]]
+        vars = self.data[self.offsets[i] : self.offsets[i + 1]]
         return vars
 
     def concat(*genos: "SparseSomaticGenotypes") -> "SparseSomaticGenotypes":
@@ -489,7 +462,7 @@ class SparseSomaticGenotypes:
             raise ValueError("All genotypes must have the same number of samples.")
 
         total_n_regions = sum(g.n_regions for g in genos)
-        variant_idxs = np.concatenate([g.variant_idxs for g in genos])
+        variant_idxs = np.concatenate([g.data for g in genos])
         offsets = _lengths_to_offsets(
             np.concatenate([np.diff(g.offsets) for g in genos])
         )
@@ -497,7 +470,7 @@ class SparseSomaticGenotypes:
         ccfs = np.concatenate([g.ccfs for g in genos if g.ccfs is not None])
 
         return SparseSomaticGenotypes(
-            variant_idxs=variant_idxs,
+            data=variant_idxs,
             offsets=offsets,
             n_regions=total_n_regions,
             n_samples=genos[0].n_samples,
@@ -539,7 +512,7 @@ class SparseSomaticGenotypes:
         # (s v) -> region/variant-major, flattened
         ccfs = ccfs.T[keep.T]
         return cls(
-            variant_idxs=variant_idxs,
+            data=variant_idxs,
             ccfs=ccfs,
             offsets=sparse_offsets,
             n_regions=n_regions,
@@ -603,7 +576,7 @@ class SparseSomaticGenotypes:
         # (s v) -> region/variant-major, flattened
         ccfs = ccfs.T[keep.T]
         sparse_genos = cls(
-            variant_idxs=variant_idxs,
+            data=variant_idxs,
             ccfs=ccfs,
             offsets=sparse_offsets,
             n_regions=n_regions,
@@ -696,12 +669,12 @@ def get_diffs_sparse(
     geno_offset_idxs: NDArray[np.intp],
     geno_v_idxs: NDArray[np.int32],
     geno_offsets: NDArray[np.int64],
-    size_diffs: NDArray[np.int32],
+    ilens: NDArray[np.int32],
     keep: Optional[NDArray[np.bool_]] = None,
     keep_offsets: Optional[NDArray[np.int64]] = None,
-    starts: Optional[NDArray[np.int32]] = None,
-    ends: Optional[NDArray[np.int32]] = None,
-    positions: Optional[NDArray[np.int32]] = None,
+    q_starts: Optional[NDArray[np.int32]] = None,
+    q_ends: Optional[NDArray[np.int32]] = None,
+    v_starts: Optional[NDArray[np.int32]] = None,
 ):
     """Get difference in length wrt reference genome for given genotypes.
 
@@ -715,17 +688,17 @@ def get_diffs_sparse(
         Shape = (variants*samples*ploidy) Sparse genotypes i.e. variant indices for ALT genotypes.
     geno_offsets : NDArray[np.int32]
         Shape = (regions*samples*ploidy + 1) Offsets into sparse genotypes.
-    size_diffs : NDArray[np.int32]
+    ilens : NDArray[np.int32]
         Shape = (total_variants) Size of all unique variants.
     keep : Optional[NDArray[np.bool_]]
         Shape = (variants*samples*ploidy) Keep mask for genotypes.
     keep_offsets : Optional[NDArray[np.int64]]
         Shape = (regions*samples*ploidy + 1) Offsets into keep.
-    starts : Optional[NDArray[np.int32]]
+    q_starts : Optional[NDArray[np.int32]]
         Shape = (regions) Start of query regions.
-    ends : Optional[NDArray[np.int32]]
+    q_ends : Optional[NDArray[np.int32]]
         Shape = (regions) End of query regions.
-    positions : Optional[NDArray[np.int32]]
+    v_starts : Optional[NDArray[np.int32]]
         Shape = (total_variants) Positions of unique variants.
     """
     n_queries, ploidy = geno_offset_idxs.shape
@@ -733,11 +706,14 @@ def get_diffs_sparse(
     for query in nb.prange(n_queries):
         for hap in nb.prange(ploidy):
             o_idx = geno_offset_idxs[query, hap]
-            o_s, o_e = geno_offsets[o_idx], geno_offsets[o_idx + 1]
+            if geno_offsets.ndim == 1:
+                o_s, o_e = geno_offsets[o_idx], geno_offsets[o_idx + 1]
+            else:
+                o_s, o_e = geno_offsets[o_idx]
             n_variants = o_e - o_s
             if n_variants == 0:
                 diffs[query, hap] = 0
-            elif starts is not None and ends is not None and positions is not None:
+            elif q_starts is not None and q_ends is not None and v_starts is not None:
                 diffs[query, hap] = 0
                 for v in range(o_s, o_e):
                     if keep is not None and keep_offsets is not None:
@@ -747,16 +723,16 @@ def get_diffs_sparse(
                             continue
 
                     v_idx: int = geno_v_idxs[v]
-                    v_pos = positions[v_idx]
-                    v_diff = size_diffs[v_idx]
+                    v_start = v_starts[v_idx]
+                    v_ilen = ilens[v_idx]
                     # +1 assumes atomized variants
-                    v_end = v_pos - min(0, v_diff) + 1
+                    v_end = v_start - min(0, v_ilen) + 1
 
-                    if v_end <= starts[query]:
+                    if v_end <= q_starts[query]:
                         # variant doesn't span region
                         continue
 
-                    if v_pos >= ends[query]:
+                    if v_start >= q_ends[query]:
                         # variants are sorted by position so this variant and everything
                         # after will be outside the region
                         break
@@ -767,20 +743,21 @@ def get_diffs_sparse(
                     # DEL r - s - e - - : +max(0, 0 - 2) -> -1 + 0 = -1
                     # where r is region start, s is variant start, e is variant end (exclusive)
                     # count the "-" to get ilen
-                    if v_diff < 0:
-                        v_diff += max(0, starts[query] - v_pos)
+                    # but also atomic deletions include 1 bp of ref so add it back (- 1)
+                    if v_ilen < 0:
+                        v_ilen += max(0, q_starts[query] - v_start - 1)
                     # deletion may end after region
-                    v_diff += max(0, v_end - ends[query])
+                    v_ilen += max(0, v_end - q_ends[query])
 
-                    diffs[query, hap] += v_diff
+                    diffs[query, hap] += v_ilen
             elif keep is not None and keep_offsets is not None:
                 v_idxs = geno_v_idxs[o_s:o_e]
                 k_idx = query * ploidy + hap
                 qh_keep = keep[keep_offsets[k_idx] : keep_offsets[k_idx + 1]]
                 v_idxs = v_idxs[qh_keep]
-                diffs[query, hap] = size_diffs[v_idxs].sum()
+                diffs[query, hap] = ilens[v_idxs].sum()
             else:
-                diffs[query, hap] = size_diffs[geno_v_idxs[o_s:o_e]].sum()
+                diffs[query, hap] = ilens[geno_v_idxs[o_s:o_e]].sum()
     return diffs
 
 
@@ -948,7 +925,11 @@ def reconstruct_haplotype_from_sparse(
     annot_ref_pos: Optional[NDArray[np.int32]]
         Shape = (out_length) Reference positions for annotations
     """
-    _variant_idxs = geno_v_idxs[geno_offsets[offset_idx] : geno_offsets[offset_idx + 1]]
+    if geno_offsets.ndim == 1:
+        o_s, o_e = geno_offsets[offset_idx], geno_offsets[offset_idx + 1]
+    else:
+        o_s, o_e = geno_offsets[offset_idx]
+    _variant_idxs = geno_v_idxs[o_s : o_e]
     length = len(out)
     n_variants = len(_variant_idxs)
 
@@ -1133,7 +1114,10 @@ def choose_unphased_variants(
     for query in nb.prange(n_regions):
         for hap in range(ploidy):
             o_idx = geno_offset_idxs[query, hap]
-            o_s, o_e = geno_offsets[o_idx], geno_offsets[o_idx + 1]
+            if geno_offset_idxs.ndim == 1:
+                o_s, o_e = geno_offsets[o_idx], geno_offsets[o_idx + 1]
+            else:
+                o_s, o_e = geno_offsets[o_idx]
             lengths[query, hap] = o_e - o_s
     keep_offsets = np.empty(n_regions * ploidy + 1, np.int64)
     keep_offsets[0] = 0
@@ -1150,7 +1134,10 @@ def choose_unphased_variants(
         ref_end: int = ends[query]
         for hap in nb.prange(ploidy):
             o_idx = geno_offset_idxs[query, hap]
-            o_s, o_e = geno_offsets[o_idx], geno_offsets[o_idx + 1]
+            if geno_offset_idxs.ndim == 1:
+                o_s, o_e = geno_offsets[o_idx], geno_offsets[o_idx + 1]
+            else:
+                o_s, o_e = geno_offsets[o_idx]
             qh_genos = geno_v_idxs[o_s:o_e]
             qh_ccfs = ccfs[o_s:o_e]